fix(sales): ROUND 인자를 numeric으로 캐스팅 — PostgreSQL float ROUND 미지원

+++ fastapi_backend/app/routes/sales.py
@@ -93,7 +93,7 @@ async def list_sales(
             )::float                                AS net_profit,
             COALESCE(
                 ROUND(
-                    COALESCE(o.rented_units, 0)::float / NULLIF(b.total_units, 0) * 100,
+                    (COALESCE(o.rented_units, 0)::numeric / NULLIF(b.total_units, 0) * 100),
                     1
                 ),
                 0
